DHCP settings and SCPI control
- basic SCPI control
- base for measuring methods
- add DHCP settings

+++ smartpower3/meas_channel.h
@@ -7,15 +7,15 @@ class MeasChannel
 {
 public:
 	MeasChannel (Microchip_PAC193x *pac, int channel_number);
-	void update (float V, float A);  // Update externally
+	void update (uint16_t V, uint16_t A);  // Update externally
 	uint16_t V (void);
 	uint16_t A (int interval);
 	uint16_t W (int interval);
 private:
 	Microchip_PAC193x *pac;  // Pointer to PAC
 	int channel_number;  // Channel number (0,1,2)
-	float voltage;
-	float current_buffer[1000];  // Circular current samplex buffer
+	uint16_t voltage;
+	uint16_t current_buffer[1000];  // Circular current samplex buffer
 	int current_buffer_count = 0;
 };
 
@@ -23,7 +23,7 @@ private:
 class MeasChannels
 {
 public:
-	MeasChannels (Microchip_PAC193x *mc, MeasChannel *Ch0, MeasChannel *Ch1, MeasChannel *Ch2, float rsense);
+	MeasChannels (Microchip_PAC193x *mc, MeasChannel *Ch0, MeasChannel *Ch1, MeasChannel *Ch2, uint16_t rsense);
 	void sample ();
 	MeasChannel* getChannel(uint8_t channel_number);
 private:
@@ -31,7 +31,7 @@ private:
 	MeasChannel *channel0;
 	MeasChannel *channel1;
 	MeasChannel *channel2;
-	float rsense;
+	uint16_t rsense;
 };
 
 #endif
